Extra functionality added and a bugfix or two - initial contribution by DC
*Hide avatar option for home, mentions and messages timeline.  Requires
app restart to banner is displayed notifying user.
*Shield direct message notification to just show that a new tweet has
arrived.  Clicking on banner just returns to home timeline (not jumping
to direct message). Option set in preferences.
*Ability to delete a received DM.  Option set in preferences.
*Bugfix for Welcome screen user names (previously flowed over right
edge).
*Added received/sent icon next to user name in DM panel to help define
direction of message (especially useful when avatars are hidden).
*Star symbol is placed next to username if tweet is favourited. Auto
update of timeline to display is set with refreshOnPost preference, and
updates on close of toaster (x or back-gesture).
*Added some bits to allow for different panel layout
*Changed id in appinfo.json to not conflict with master projectmacaw
app, but kept version to show inherited version
*Added extra icons (arrow_left.png and arrow_right.png).  Not used as
using unicode character from font.  main-assistant.js is now unicode
instead of ascii as a result