Inline images added plus more
Preference option allows inline images to be viewed in timelines and
details view.  In details view tapping on image loads preview or
webpage.
Updated the way that the favourite star and DM direction arrows are
displayed.
Removed "via" mention in details view of DMs.
Added time of tweet (or date if greater than 24hrs) in details view.