More inline thumbs
Added display of inline thumbs of direct .jpg, .jpeg, .png and .gif
images.  Also, emoji and inline thumbs are now displayed in search
results.
Option was added to hide inline thumbs in search timelines to give the
user the option to hide "unexpected" images from popping up.  At the
moment its just a toggle, but I might add it to drop down list if there
is a nice way to have that option.