Workaround for annoying bug with convo-scrollers
Before the tweak, you couldn't select the first item unless you were on
a TP and selected beside it in the location of the next panel.
Can't work out what is causing the prevention of tapping the first
entry, but by adding the pad, it pushes down the item so it clears the
obstruction.